+++ dxf_checker/__init__.py
@@ -2,6 +2,6 @@
 CheckMateLL DXF Checker - A tool for validating DXF segment integrity
 """
 
-__version__ = "2.0.0" # {bumpver}
+__version__ = "2.1.0" # {bumpver}
 __author__ = "Juliya Lehka"
 __email__ = "julilya.legkaya@gmail.com"